Add services

+++ SolarSystemCore/SolarSystemCore.Services/PlanetService.cs
@@ -4,6 +4,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq.Expressions;
 using System.Threading.Tasks;
+using System.Linq;
 
 namespace SolarSystemCore.Services
 {
@@ -14,6 +15,8 @@ namespace SolarSystemCore.Services
         public PlanetService(IRepository<Planet> planetRepository) => this.planetRepository = planetRepository;
 
         public async Task<IEnumerable<Planet>> GetAllPlanetsAsync() => await planetRepository.GetAllAsync();
+
+        public async Task<IEnumerable<Planet>> GetAllPlanetsByStarIdAsync(int starId) => await planetRepository.FindAsync(p => p.StarId == starId);
 
         public async Task<Planet> GetPlanetAsync(int id) => await planetRepository.FirstOrDefaultAsync(p => p.Id == id);
 
@@ -28,8 +31,17 @@ namespace SolarSystemCore.Services
             throw new NullReferenceException();
         }
 
-        public async Task<int> AddPlanetsAsync(IList<Planet> planets) => await planetRepository.AddRangeAsync(planets);
+        public async Task<int> AddPlanetsAsync(IList<Planet> planets)
+        {
+            var planetsToAdd = planets.Where(p => p.Name != null || p.Name != string.Empty).ToList();
 
+            if (planetsToAdd.Count() > 0)
+            {
+                return await planetRepository.AddRangeAsync(planets);
+            }
+            throw new ArgumentException();
+        }
+        
         public async Task<int> SavePlanetAsync(Planet planet)
         {
             if (planet.Id != 0)
@@ -43,7 +55,7 @@ namespace SolarSystemCore.Services
         {
             var planet = await GetPlanetAsync(id);
 
-            if (planet != null)
+            if (planet != null && planet.Id == id)
             {
                 return await planetRepository.DeleteAsync(planet);
             }
